Moved remaining functions from DashboardUtil to DashboardService and deleted DashboardUtil.
